refactor: move internal files
