+++ packages/cli/src/api/locales.js
