+++ packages/cli/src/api/locales.test.js
